fix(graphviz): incr_width halves nodesep with C INTEGER division
class2.c:43

    static void incr_width(graph_t * g, node_t * v) {
        int width = GD_nodesep(g) / 2;
        ND_lw(v) += width;
        ND_rw(v) += width;
    }

`int width = GD_nodesep(g) / 2` truncates, and GD_nodesep is itself an int
(POINTS rounds), so an ODD nodesep loses the half: 191 sets nodesep=0.35 =>
25 => 12, where we computed 12.5. At the 18pt default both give 9, which is
why no other corpus file caught it.

Found by diffing the aux-graph CONSTRAINTS rather than the coordinates: gv's
keepout_othernodes edges around cluster_package_programs_para were len 21
(8 + 13) against our 22 (8 + 13.5). With this, every constraint touching that
cluster's ln/rn matches gv exactly.

It does not move 191's coordinates on its own — those constraints were not
binding — but it is the same class of bug PORT.md already records for
Splinesep (`GD_nodesep(g) / 4` vs `/ 4.0`), and it is wrong wherever nodesep
is odd. 779 graphviz tests green; 162 corpus files stay byte-exact.

+++ graphviz/shared/src/main/scala/org/jpablo/graphexplorer/graphviz/layout/Coord.scala
@@ -62,8 +62,13 @@ object Coord:
     points(v)
 
   /** Plain virtual node half-width — `incr_width` = 1 + nodesep/2
-    * (class2.c plain_vnode). Shared by [[XCoord]] and [[Spline]]. */
-  def virtualHalfPt(g: RGraph): Double = 1.0 + nodeSepPt(g) / 2.0
+    * (class2.c plain_vnode). Shared by [[XCoord]] and [[Spline]].
+    *
+    * `int width = GD_nodesep(g) / 2` is C INTEGER division, and GD_nodesep is
+    * itself an int (`POINTS` rounds), so an ODD nodesep truncates: 191 sets
+    * `nodesep=0.35` ⇒ 25 ⇒ 12, not 12.5. Invisible at the 18pt default (9
+    * either way), which is why no other corpus file caught it. */
+  def virtualHalfPt(g: RGraph): Double = 1.0 + (nodeSepPt(g).toInt / 2).toDouble
 
   /** `selfRightSpace(e)` (splines.c:1146): the extra RIGHT-side space a
     * self-edge needs — `SELF_EDGE_SIZE` (18) plus its label width (`dimen.y`
